refactor controllers into one package
