+++ src/main/java/com/glovo/ttglovo/controllers/appUser/AppUserInfoController.java
@@ -1,5 +1,7 @@
-package com.glovo.ttglovo.securityManagement.appuser;
+package com.glovo.ttglovo.controllers.appUser;
 
+import com.glovo.ttglovo.securityManagement.appuser.AppUser;
+import com.glovo.ttglovo.securityManagement.appuser.AppUserService;
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
